move route handlers to the route/
