+++ packages/stream-metadata/src/routes/handleMetadataRequest.ts
@@ -1,6 +1,6 @@
 import { FastifyRequest, FastifyReply } from 'fastify'
 
-import { isValidEthereumAddress } from './validators'
+import { isValidEthereumAddress } from '../validators'
 
 export function handleMetadataRequest(
 	request: FastifyRequest,
